fix animation (#60)
* fix animation

* Update update_objects.c

+++ include/course_offsets.h
@@ -15,7 +15,7 @@ typedef struct
     u32 padding;   // always zero
 } course_texture;
 
-extern u32 d_course_sherbet_land_unk_data1[];
+extern uintptr_t d_course_sherbet_land_unk_data1[];
 extern s16 d_course_sherbet_land_unk_data2[];
 extern animation_type_3_triplet d_course_sherbet_land_unk_data3[];
 extern animation_type_2 d_course_sherbet_land_unk_data4;
@@ -52,7 +52,7 @@ extern animation_type_2 *d_course_yoshi_valley_unk4[];
 extern uintptr_t d_course_yoshi_valley_unk5[];
 extern u32 d_course_yoshi_valley_unk6[];
 
-extern u32 d_course_sherbet_land_unk_data1[];
+extern uintptr_t d_course_sherbet_land_unk_data1[];
 extern s16 d_course_sherbet_land_unk_data2[];
 extern animation_type_3_triplet d_course_sherbet_land_unk_data3[];
 extern animation_type_2 d_course_sherbet_land_unk_data4;
